Finish 4.21

+++ controllers/blogs.js
@@ -23,15 +23,6 @@ blogsRouter.get('/:id', (request, response, next) => {
     .catch(error => next(error))
 })
 
-// const getTokenFrom = request => {
-//   const authorization = request.get('authorization')
-
-//   if (authorization && authorization.toLowerCase().startsWith('bearer')) {
-//     return authorization.substring(7)
-//   }
-//   return null
-// }
-
 blogsRouter.post('/', async (request, response) => {
   const body = request.body
   
@@ -58,8 +49,25 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndRemove(request.params.id)
-  response.status(204).end()
+  // const body = request.body
+  // console.log('request body', body)
+
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken) {
+    return response.status(401).json({ error: 'token missing or invalid' })
+  }
+  const currUser = await User.findById(decodedToken.id)
+
+  const blog = await Blog.findById(request.params.id)
+
+  console.log('blog to delete', blog)
+
+  if (blog.user.toString() === currUser._id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(403).json({ error: 'Not authorized to delete other users\' posts'})
+  }
 })
 
 blogsRouter.put('/:id', async (request, response) => {
